Move docker related stuff to docker NS
