+++ .github/workflows/ci.yml
@@ -79,6 +79,7 @@ jobs:
                     <?php
 
                     use Castor\Attribute\AsTask;
+                    use function docker\docker_compose_run;
 
                     #[AsTask()]
                     function test()
